dashboard controller added

+++ src/PersonalFinanceManager.Application/DTOs/Account/AccountDto.cs
@@ -7,7 +7,7 @@ public class AccountDto
     public string Name { get; set; } = string.Empty;
     public AccountType Type { get; set; }
     public string Institution { get; set; } = string.Empty;
-    public decimal CurrentBalance { get; set; }
+    public double CurrentBalance { get; set; }
     public string Currency { get; set; } = "BDT";
     public bool IsActive { get; set; }
     public DateTime? UpdatedAt { get; set; }
