+++ src/PersonalFinanceManager.Application/DTOs/Account/CreateAccountDto.cs
@@ -7,7 +7,7 @@ public class CreateAccountDto
     public string Name { get; set; } = string.Empty;
     public AccountType Type { get; set; }
     public string Institution { get; set; } = string.Empty;
-    public decimal CurrentBalance { get; set; }
+    public double CurrentBalance { get; set; }
     public string Currency { get; set; } = "BDT";
     public string? Description { get; set; }
     public bool IncludeInNetWorth { get; set; } = true;
